refactor: move local state to plugin
